requirements.txt
transformers==4.19.2
diffusers==0.11.1
basicsr==1.4.2
einops==0.6.0
omegaconf==2.3.0
pytorch_lightning==1.5.9
opencv-python
open-clip-torch
safetensors
